% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQE-package.R
\docType{package}
\name{SQE-package}
\alias{SQE}
\alias{SQE-package}
\title{SQE: Sorghum QTL effects (in BCNAM populations - draft)}
\description{
Application to get parental QTL allelic effects estimated in different (multi-environment) sorghum experiments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vincentgarin/SQE}
  \item Report bugs at \url{https://github.com/vincentgarin/SQE/issues}
}

}
\author{
\strong{Maintainer}: Vincent Garin \email{vincent.garin6@gmail.com} (\href{https://orcid.org/0000-0002-5571-1841}{ORCID})

}
\keyword{internal}
